[wpiutil] Upgrade to fmt 9.0.0 (#4337)
fmt removed fmt::make_args_checked since it's no longer needed for
constexpr format string checks.

fmt deprecated implicit conversions from enums to integers in format
arguments, so I added explicit static casts.

+++ cscore/src/main/native/cpp/MjpegServerImpl.cpp
@@ -495,7 +495,7 @@ void MjpegServerImpl::ConnThread::SendJSON(wpi::raw_ostream& os,
     auto kind = source.GetPropertyKind(prop);
     fmt::print(os, "\n\"name\": \"{}\"", name);
     fmt::print(os, ",\n\"id\": \"{}\"", prop);
-    fmt::print(os, ",\n\"type\": \"{}\"", kind);
+    fmt::print(os, ",\n\"type\": \"{}\"", static_cast<int>(kind));
     fmt::print(os, ",\n\"min\": \"{}\"", source.GetPropertyMin(prop, &status));
     fmt::print(os, ",\n\"max\": \"{}\"", source.GetPropertyMax(prop, &status));
     fmt::print(os, ",\n\"step\": \"{}\"",
